move s3_uri
