+++ be/test/fs/fs_s3_test.cpp
@@ -23,8 +23,8 @@
 #include <fstream>
 
 #include "common/config.h"
-#include "common/s3_uri.h"
 #include "fs/fs_s3.h"
+#include "fs/s3_uri.h"
 #include "gutil/strings/join.h"
 #include "testutil/assert.h"
 #include "util/uid_util.h"
